fix(cb2-4081): refactor so connections call is only made once and not for each account
