+++ tests/unit/dynamicsWebApi.test.ts
@@ -3,14 +3,11 @@ import { of } from 'rxjs';
 import { AxiosResponse, AxiosError } from 'axios';
 import {
   onRejected,
-  mapToDynamoTestStation,
   getModifiedTestStations,
-  getTestStationEntities,
   getReportRecipientEmails,
 } from '../../src/crm/dynamicsWebApi';
-// import { DynamoTestStation } from '../../src/crm/DynamoTestStation';
 import { DynamicsTestStation } from '../../src/crm/DynamicsTestStation';
-import { DynamoTestStation } from '../../src/crm/DynamoTestStation';
+import { DynamicsConnection } from '../../src/crm/DynamicsConnection';
 
 jest.mock('../../src/crm/getToken', () => ({
   getToken: jest.fn().mockResolvedValueOnce({ value: 'MOCKED_BEARER_TOKEN' }),
@@ -18,26 +15,6 @@ jest.mock('../../src/crm/getToken', () => ({
 jest.mock('../../src/utils/index');
 
 describe('dynamicsWebApi', () => {
-  const MOCK_BAD_ACCOUNTS_DATA: DynamicsTestStation = {
-    '@odata.etag': 'string',
-    accountid: '1234',
-    address1_composite: 'string',
-    address1_line1: 'Address 1',
-    address1_line2: 'Address 2',
-    telephone1: 'string',
-    emailaddress1: 'string',
-    dvsa_openingtimes: 'string',
-    address1_longitude: 'string',
-    address1_latitude: 'string',
-    name: 'string',
-    dvsa_premisecodes: 'string',
-    address1_postalcode: 'string',
-    dvsa_accountstatus: 147160001,
-    address1_city: 'string',
-    dvsa_testfacilitytype: 1471600,
-    modifiedon: '',
-  };
-
   const MOCK_GETACCOUNTS_RETURN: DynamicsTestStation[] = [
     {
       '@odata.etag': 'string',
@@ -76,6 +53,61 @@ describe('dynamicsWebApi', () => {
       address1_city: 'string',
       dvsa_testfacilitytype: 147160000,
       modifiedon: '',
+    },
+  ];
+
+  const MOCK_GETCONNECTIONS_RESPONSE: DynamicsConnection[] = [
+    {
+      '@odata.etag': 'string',
+      _record2id_value: 'string',
+      connectionid: 'string',
+      record1id_account: {
+        accountid: 'string',
+        accountnumber: '00000',
+      },
+      record2id_contact: {
+        emailaddress1: 'email@email.co.uk',
+        contactid: 'string',
+      },
+    },
+    {
+      '@odata.etag': 'string',
+      _record2id_value: 'string',
+      connectionid: 'string',
+      record1id_account: {
+        accountid: 'string',
+        accountnumber: '00000',
+      },
+      record2id_contact: {
+        emailaddress1: 'email2@email.co.uk',
+        contactid: 'string',
+      },
+    },
+    {
+      '@odata.etag': 'string',
+      _record2id_value: 'string',
+      connectionid: 'string',
+      record1id_account: {
+        accountid: 'string',
+        accountnumber: '00000',
+      },
+      record2id_contact: {
+        emailaddress1: 'email3@email.co.uk',
+        contactid: 'string',
+      },
+    },
+    {
+      '@odata.etag': 'string',
+      _record2id_value: 'string',
+      connectionid: 'string',
+      record1id_account: {
+        accountid: 'string',
+        accountnumber: '00000',
+      },
+      record2id_contact: {
+        emailaddress1: 'email@email.co.uk',
+        contactid: 'string',
+      },
     },
   ];
 
@@ -130,88 +162,62 @@ describe('dynamicsWebApi', () => {
     config: {},
   };
 
-  const MOCK_CONNECTIONS_RESPONSE: AxiosResponse[] = [
-    {
-      data: {
-        value: [
-          {
-            '@odata.etag': 'string',
-            _record2id_value: 'string',
-            connectionid: 'string',
-            record1id_account: {
-              accountid: 'string',
-              accountnumber: '00000',
-            },
-            record2id_contact: {
-              emailaddress1: 'email@email.co.uk',
-              contactid: 'string',
-            },
-          },
-          {
-            '@odata.etag': 'string',
-            _record2id_value: 'string',
-            connectionid: 'string',
-            record1id_account: {
-              accountid: 'string',
-              accountnumber: '00000',
-            },
-            record2id_contact: {
-              emailaddress1: 'email2@email.co.uk',
-              contactid: 'string',
-            },
-          },
-          {
-            '@odata.etag': 'string',
-            _record2id_value: 'string',
-            connectionid: 'string',
-            record1id_account: {
-              accountid: 'string',
-              accountnumber: '00000',
-            },
-            record2id_contact: {
-              emailaddress1: 'email3@email.co.uk',
-              contactid: 'string',
-            },
-          },
-        ],
-      },
-      status: 200,
-      statusText: 'Ok',
-      headers: {
-        Authorization: '',
-      },
-      config: {},
-    },
-    {
-      data: {
-        value: [
-          {
-            '@odata.etag': 'string',
-            _record2id_value: 'string',
-            connectionid: 'string',
-            record1id_account: {
-              accountid: 'string',
-              accountnumber: '00000',
-            },
-            record2id_contact: {
-              emailaddress1: 'email@email.co.uk',
-              contactid: 'string',
-            },
-          },
-        ],
-      },
-      status: 200,
-      statusText: 'Ok',
-      headers: {
-        Authorization: '',
-      },
-      config: {},
-    },
-  ];
-
-  const MOCK_CONNECTIONS_NO_EMAILS_RESPONSE: AxiosResponse = {
+  const MOCK_CONNECTIONS_RESPONSE: AxiosResponse = {
     data: {
-      value: [],
+      value: [
+        {
+          '@odata.etag': 'string',
+          _record2id_value: 'string',
+          connectionid: 'string',
+          record1id_account: {
+            accountid: 'string',
+            accountnumber: '00000',
+          },
+          record2id_contact: {
+            emailaddress1: 'email@email.co.uk',
+            contactid: 'string',
+          },
+        },
+        {
+          '@odata.etag': 'string',
+          _record2id_value: 'string',
+          connectionid: 'string',
+          record1id_account: {
+            accountid: 'string',
+            accountnumber: '00000',
+          },
+          record2id_contact: {
+            emailaddress1: 'email2@email.co.uk',
+            contactid: 'string',
+          },
+        },
+        {
+          '@odata.etag': 'string',
+          _record2id_value: 'string',
+          connectionid: 'string',
+          record1id_account: {
+            accountid: 'string',
+            accountnumber: '00000',
+          },
+          record2id_contact: {
+            emailaddress1: 'email3@email.co.uk',
+            contactid: 'string',
+          },
+        },
+        {
+          '@odata.etag': 'string',
+          _record2id_value: 'string',
+          connectionid: 'string',
+          record1id_account: {
+            accountid: 'string',
+            accountnumber: '00000',
+          },
+          record2id_contact: {
+            emailaddress1: 'email@email.co.uk',
+            contactid: 'string',
+          },
+        },
+      ],
     },
     status: 200,
     statusText: 'Ok',
@@ -219,42 +225,7 @@ describe('dynamicsWebApi', () => {
       Authorization: '',
     },
     config: {},
-  }
-
-  const MOCK_RESULT: DynamoTestStation[] = [
-    {
-      testStationId: 'string',
-      testStationAccessNotes: null,
-      testStationAddress: 'Address 1',
-      testStationContactNumber: 'string',
-      testStationEmails: ['email@email.co.uk', 'email2@email.co.uk', 'email3@email.co.uk'],
-      testStationGeneralNotes: 'string',
-      testStationLongitude: 'string',
-      testStationLatitude: 'string',
-      testStationName: 'string',
-      testStationPNumber: 'P601',
-      testStationPostcode: 'string',
-      testStationStatus: 'active',
-      testStationTown: 'string',
-      testStationType: 'atf',
-    },
-    {
-      testStationId: 'string',
-      testStationAccessNotes: null,
-      testStationAddress: 'Address 1, Address 2',
-      testStationContactNumber: 'string',
-      testStationEmails: ['email@email.co.uk'],
-      testStationGeneralNotes: null,
-      testStationLongitude: 'string',
-      testStationLatitude: 'string',
-      testStationName: 'string',
-      testStationPNumber: 'P602',
-      testStationPostcode: 'string',
-      testStationStatus: 'active',
-      testStationTown: 'string',
-      testStationType: 'atf',
-    },
-  ];
+  };
 
   const MOCK_FAILURE: AxiosError = {
     name: '',
@@ -281,20 +252,11 @@ describe('dynamicsWebApi', () => {
     );
   });
 
-  test('GIVEN mock axios odata succesful response with invalid testStationPNumber WHEN called THEN throws an error', () => {
-    function throwAnError() {
-      mapToDynamoTestStation(MOCK_BAD_ACCOUNTS_DATA);
-    }
-    expect(throwAnError).toThrowError(
-      new Error('Invalid enum value provided for test station type field: 1471600 for test station: 1234'),
-    );
-  });
-
-  test('GIVEN mock axios odata successful response from connections table WHEN called THEN returns array of emails', async () => {
-    const EMAILS = ['email@email.co.uk', 'email2@email.co.uk', 'email3@email.co.uk'];
-    axios.get = jest.fn().mockReturnValueOnce(of(MOCK_CONNECTIONS_RESPONSE[0]));
-    const result = await getReportRecipientEmails('');
-    expect(result).toEqual(EMAILS);
+  test('GIVEN mock axios odata successful response from connections table WHEN called THEN returns array of DynamicsConnection objects', async () => {
+    axios.get = jest.fn().mockReturnValueOnce(of(MOCK_CONNECTIONS_RESPONSE));
+    const result = await getReportRecipientEmails();
+    expect(result).toHaveLength(4);
+    expect(result).toEqual(MOCK_GETCONNECTIONS_RESPONSE);
   });
 
   test('GIVEN mock axios odata succesful response from accounts table WHEN called THEN returns array of DynamicsTestStation objects', async () => {
@@ -303,31 +265,4 @@ describe('dynamicsWebApi', () => {
     expect(result).toHaveLength(2);
     expect(result).toEqual(MOCK_GETACCOUNTS_RETURN);
   });
-
-  test('GIVEN mock axios odata succesful response from accounts & connections table WHEN getTestStationEntities called THEN returns an array of DynamoTestStation objects containing array of email addresses', async () => {
-    axios.get = jest
-      .fn()
-      .mockReturnValueOnce(of(MOCK_ACCOUNTS_RESPONSE))
-      .mockReturnValueOnce(of(MOCK_CONNECTIONS_RESPONSE[0]))
-      .mockReturnValueOnce(of(MOCK_CONNECTIONS_RESPONSE[1]));
-
-    const result = await getTestStationEntities('');
-    expect(result).toHaveLength(2);
-    expect(result).toEqual(MOCK_RESULT);
-    expect(result[0].testStationEmails).toHaveLength(3);
-    expect(result[1].testStationEmails).toHaveLength(1);
-  });
-
-  test('GIVEN succesful response from connections table WHEN there are no emails present THEN return DynamoTestStation object with empty array for emails', async () => {
-    axios.get = jest
-      .fn()
-      .mockReturnValueOnce(of(MOCK_ACCOUNTS_RESPONSE))
-      .mockReturnValueOnce(of(MOCK_CONNECTIONS_NO_EMAILS_RESPONSE))
-      .mockReturnValueOnce(of(MOCK_CONNECTIONS_RESPONSE[1]));
-
-    const result = await getTestStationEntities('');
-    expect(result).toHaveLength(2);
-    expect(result[0].testStationEmails).toHaveLength(0);
-    expect(result[1].testStationEmails).toHaveLength(1);
-  })
 });
